\name{read.ancestrydna.web}
\alias{read.ancestrydna.web}
\title{Read AncestryDNA file from web}
\usage{
  read.ancestrydna.web(url)
}
\arguments{
  \item{url}{URL web address to raw DNA zip file downloaded from AncestryDNA}
}

\examples{
  mydna <- read.ancestrydna.web("http://example.com/dna-data.zip")
}

